+++ Assets/Scripts/InputHandlerDigitalAxis1D.cs
@@ -1,6 +1,5 @@
 using UnityEngine;
 using UnityEngine.InputSystem;
-using System.Collections;
 using Unity.VisualScripting;
 
 public class InputHandlerDigitalAxis1D : MonoBehaviour
@@ -11,8 +10,8 @@ public class InputHandlerDigitalAxis1D : MonoBehaviour
 
     public void Trigger(InputAction.CallbackContext context)
     {
-        args.playerID = (short) playerInput.playerIndex;
-        args.axis = (short) Mathf.RoundToInt(context.ReadValue<float>());
+        args.playerID = playerInput.playerIndex;
+        args.axis = Mathf.RoundToInt(context.ReadValue<float>());
         
         args.action = action;
         EventBus.Trigger(InputAxis1DEvent.Hook, args);
